upgrade Enemy Controller and Animation, logic

+++ Assets/Scripts/Interface/IHitable.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public interface IHitable
 {
     void takeDamage();
-    void takeDamage(int damage);
-    void takeDamage(int damage, string damageDealer);
+    void takeDamage(float damage);
+    void takeDamage(float damage, string damageDealer);
 }
